refactor(api): optimize OpenLibraryAPI for performance and lower latency
- Replaced per-request AsyncClient instantiation with a single, reusable client for connection pooling.
- Removed redundant argument handling in fetch_data to simplify the code path.
- Reduced overhead and improved response latency by streamlining error handling.
